+++ spring-ai-spring-boot-testcontainers/src/test/java/org/springframework/ai/testcontainers/service/connection/chroma/ChromaWithTokenContainerConnectionDetailsFactoryIT.java
@@ -45,15 +45,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 @Testcontainers
 @TestPropertySource(properties = { "spring.ai.vectorstore.chroma.collectionName=TestCollection",
 		"spring.ai.vectorstore.chroma.initialize-schema=true" })
-class ChromaWithToken2ContainerConnectionDetailsFactoryTest {
+class ChromaWithTokenContainerConnectionDetailsFactoryIT {
 
 	@Container
 	@ServiceConnection
 	static ChromaDBContainer chroma = new ChromaDBContainer(ChromaImage.DEFAULT_IMAGE)
-		.withEnv("CHROMA_SERVER_AUTH_CREDENTIALS", "token")
-		.withEnv("CHROMA_SERVER_AUTH_CREDENTIALS_PROVIDER",
-				"chromadb.auth.token.TokenConfigServerAuthCredentialsProvider")
-		.withEnv("CHROMA_SERVER_AUTH_PROVIDER", "chromadb.auth.token.TokenAuthServerProvider");
+		.withEnv("CHROMA_SERVER_AUTHN_CREDENTIALS", "token")
+		.withEnv("CHROMA_SERVER_AUTHN_PROVIDER", "chromadb.auth.token_authn.TokenAuthenticationServerProvider");
 
 	@Autowired
 	private VectorStore vectorStore;
