+++ spring-ai-spring-boot-testcontainers/src/test/java/org/springframework/ai/testcontainers/service/connection/qdrant/QdrantContainerWithApiKeyConnectionDetailsFactoryIT.java
@@ -14,19 +14,19 @@
  * limitations under the License.
  */
 
-package org.springframework.ai.testcontainers.service.connection.typesense;
+package org.springframework.ai.testcontainers.service.connection.qdrant;
 
-import java.time.Duration;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Map;
 
 import org.junit.jupiter.api.Test;
-import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.qdrant.QdrantContainer;
 
-import org.springframework.ai.ResourceUtils;
-import org.springframework.ai.autoconfigure.vectorstore.typesense.TypesenseVectorStoreAutoConfiguration;
+import org.springframework.ai.autoconfigure.vectorstore.qdrant.QdrantVectorStoreAutoConfiguration;
 import org.springframework.ai.document.Document;
 import org.springframework.ai.embedding.EmbeddingModel;
 import org.springframework.ai.transformers.TransformersEmbeddingModel;
@@ -37,57 +37,60 @@ import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.io.DefaultResourceLoader;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringJUnitConfig
-@TestPropertySource(properties = { "spring.ai.vectorstore.typesense.embeddingDimension=384",
-		"spring.ai.vectorstore.typesense.initialize-schema=true",
-		"spring.ai.vectorstore.typesense.collectionName=myTestCollection" })
 @Testcontainers
-class TypesenseContainerConnectionDetailsFactoryTest {
+@TestPropertySource(properties = { "spring.ai.vectorstore.qdrant.collectionName=test_collection",
+		"spring.ai.vectorstore.qdrant.initialize-schema=true" })
+public class QdrantContainerWithApiKeyConnectionDetailsFactoryIT {
 
 	@Container
 	@ServiceConnection
-	private static final GenericContainer<?> typesense = new GenericContainer<>(TypesenseImage.DEFAULT_IMAGE)
-		.withExposedPorts(8108)
-		.withCommand("--data-dir", "/tmp", "--enable-cors")
-		.withEnv("TYPESENSE_API_KEY", "secret")
-		.withStartupTimeout(Duration.ofSeconds(100));
+	static QdrantContainer qdrantContainer = new QdrantContainer(QdrantImage.DEFAULT_IMAGE).withApiKey("test_api_key");
 
 	List<Document> documents = List.of(
-			new Document(ResourceUtils.getText("classpath:/test/data/spring.ai.txt"), Map.of("spring", "great")),
-			new Document(ResourceUtils.getText("classpath:/test/data/time.shelter.txt")), new Document(
-					ResourceUtils.getText("classpath:/test/data/great.depression.txt"), Map.of("depression", "bad")));
+			new Document(getText("classpath:/test/data/spring.ai.txt"), Map.of("spring", "great")),
+			new Document(getText("classpath:/test/data/time.shelter.txt")),
+			new Document(getText("classpath:/test/data/great.depression.txt"), Map.of("depression", "bad")));
 
 	@Autowired
 	private VectorStore vectorStore;
 
+	public static String getText(String uri) {
+		var resource = new DefaultResourceLoader().getResource(uri);
+		try {
+			return resource.getContentAsString(StandardCharsets.UTF_8);
+		}
+		catch (IOException e) {
+			throw new RuntimeException(e);
+		}
+	}
+
 	@Test
 	public void addAndSearch() {
-
 		this.vectorStore.add(this.documents);
 
-		List<Document> results = this.vectorStore.similaritySearch(SearchRequest.query("Spring").withTopK(1));
+		List<Document> results = this.vectorStore
+			.similaritySearch(SearchRequest.query("What is Great Depression?").withTopK(1));
 
 		assertThat(results).hasSize(1);
 		Document resultDoc = results.get(0);
-		assertThat(resultDoc.getId()).isEqualTo(this.documents.get(0).getId());
-		assertThat(resultDoc.getContent())
-			.contains("Spring AI provides abstractions that serve as the foundation for developing AI applications.");
-		assertThat(resultDoc.getMetadata()).hasSize(2);
-		assertThat(resultDoc.getMetadata()).containsKeys("spring", "distance");
+		assertThat(resultDoc.getId()).isEqualTo(this.documents.get(2).getId());
+		assertThat(resultDoc.getMetadata()).containsKeys("depression", "distance");
 
+		// Remove all documents from the store
 		this.vectorStore.delete(this.documents.stream().map(doc -> doc.getId()).toList());
-
-		results = this.vectorStore.similaritySearch(SearchRequest.query("Spring").withTopK(1));
+		results = this.vectorStore.similaritySearch(SearchRequest.query("Great Depression").withTopK(1));
 		assertThat(results).hasSize(0);
 	}
 
 	@Configuration(proxyBeanMethods = false)
-	@ImportAutoConfiguration(TypesenseVectorStoreAutoConfiguration.class)
+	@ImportAutoConfiguration(QdrantVectorStoreAutoConfiguration.class)
 	static class Config {
 
 		@Bean
